Began Implementing Dungeon Class
Began the implementation of two classes: a dungeonfloor class to hold
all information pertaining to a single floor of a dungeon, and the
dungeon class itself, which stores a list of all the dungeon floors.
This will eventually allow for pre-generation of floors, speeding up the
game at runtime and also allowing the saving and loading of levels
between sessions. Furthermore, it will allow the player to go up stairs
and revisit previous floors if they have forgotten anything.