Map Generator Fixed
The map generator now works correctly and no longer throws errors that
were previously caused by regions being too small. The algorithm is now
fully scalable and there is the potential to remove the hard room limit
and instead allow the map to be split as many times as possible.